+++ tests/test_inline.rs
@@ -30,14 +30,6 @@ fn test_single_line() {
 }
 
 #[test]
-fn test_trailing_commas() {
-    assert_snapshot!(
-        "Testing",
-        @"Testing",
-    );
-}
-
-#[test]
 fn test_unnamed_single_line() {
     assert_snapshot!("Testing");
     assert_snapshot!("Testing-2");